Store the lastid file in the users home
This patch uses Dir.home and therefore assumes that you have Ruby1.9.
Don't know if we impose this on the user so far. Also it creates a .dp2
directory in the users home and puts the lastid in there. Here we assume
that there might be more files coming, e.g. config.yml